feat: enhance profile update logic to invalidate refresh tokens on password change

+++ src/modules/profile/profile.service.js
@@ -83,7 +83,7 @@ class ProfileService {
 
       // invalidate all existing refresh tokens for the user to force logout
       // from all devices if password is changed
-      if (data.password || newPassword.length) {
+      if (data.password && newPassword.length) {
         await authRepository.deleteRefreshTokensByUserId(userId, tx)
       }
